working save

+++ src/main/java/com/example/demo/repository/EscalationRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EscalationRepository extends JpaRepository<Escalation, Long> { }
+public interface EscalationRepository extends JpaRepository<Escalation, Integer> { }
